fix: import 위치 변경, default변경, 불필요한 value제거 move: chat-room, chat 폴더 이동
