+++ src/components/overlay/chatting/components/index.js
@@ -1,0 +1,4 @@
+import Chat from "./chat";
+import ChatRoom from "./chat-room";
+
+export { Chat, ChatRoom };
